test case

+++ pkg/log/encoder.go
@@ -29,5 +29,5 @@ func timeEncoder(t time.Time, enc zapcore.PrimitiveArrayEncoder) {
 }
 
 func milliSecondsDurationEncoder(d time.Duration, enc zapcore.PrimitiveArrayEncoder) {
-	enc.AppendFloat64(float64(d) / float64(time.Millisecond))
+	enc.AppendInt64(int64(d / time.Millisecond))
 }
